server/requirements.txt
fastapi
uvicorn
langchain-ollama
langchain-community
chromadb
python-dotenv
langchain-google-genai
langchain-chroma
langchain-groq